refactor: move SpanDelegates to Light.GuardClauses.ExceptionFactory namespace
